refactor: :recycle: move preference stuffs to preference folder
