+++ lib/common/config/theme.dart
@@ -30,10 +30,10 @@ const MaterialColor primary = MaterialColor(0xFF0AA1DD, <int, Color>{
   50: Color(0xFF0AA1DD),
   100: Color(0xFF0AA1DD),
   200: Color(0xFF0AA1DD),
-  300: Color(0xFFE1F5FD),
-  400: Color(0xFF0AA1DD),
+  300: Color(0xFF0AA1DD),
+  400: Color(0xFFE1F5FD),
   500: Color(0xFF0AA1DD),
-  600: Color(0xFF0AA1DD),
+  600: Color(0xFF184B5F),
   700: Color(0xFF0AA1DD),
   800: Color(0xFF0AA1DD),
   900: Color(0xFF0AA1DD),
